refacto(use-cases): extract shared PromptOnlyForwardingUseCase to deduplicate ocr/rerank/embeddings
