+++ api/tests/unit/use_case/embeddings/test_createembeddingsusecase.py
@@ -190,7 +190,7 @@ def sample_embeddings():
 @pytest.fixture
 def mock_embeddings_latency_120ms():
     with patch(
-        "api.use_cases.embeddings._createembeddingsusecase.time.perf_counter",
+        "api.use_cases._forwarding.time.perf_counter",
         side_effect=[0, 0.12],
     ):
         yield
